chore: 파일 정리, 리팩터링, 로딩 상태 처리
