+++ src/components/molecules/index.ts
@@ -4,6 +4,7 @@ export { CountEmoji } from './countEmoji';
 export { EmojiGroup } from './emojiGroup';
 export { InfiniteScroller } from './infiniteScroll';
 export { LimitedLengthInput } from './limitedLengthInput';
+export { Loading } from './loading';
 export { LoginIconSet } from './loginIconSet';
 export { SelectableCardList } from './selectableCardList';
 export { Skeleton } from './skeleton';
