+++ src/features/comment/BottomSheet/atom.ts
@@ -1,0 +1,3 @@
+import { atom } from 'jotai';
+
+export const goalIdAtom = atom(-1);
