+++ src/hooks/reactQuery/comment/useGetComment.ts
@@ -1,4 +1,4 @@
-import { useQuery } from '@tanstack/react-query';
+import { useSuspenseQuery } from '@tanstack/react-query';
 
 import { api } from '@/apis';
 
@@ -27,7 +27,7 @@ type CommenterProps = {
 };
 
 export const useGetComment = ({ goalId }: CommentRequestParams) => {
-  return useQuery<CommentResponse>({
+  return useSuspenseQuery<CommentResponse>({
     queryKey: ['comment', goalId],
     queryFn: () => api.get<CommentResponse>(`/goal/${goalId}/comment`),
     staleTime: 1000,
